Make required numeric and unconstrained, like limitation
Drops the two CHECK constraints so required behaves the same as the
existing limitation column: an optional count the database does not
validate. Bad values are a data issue, and an admin can no longer hit
a generic save error from a constraint violation.

+++ supabase/functions/get_fighter_types_with_cost.sql
@@ -36,7 +36,7 @@ RETURNS TABLE (
     attacks numeric,
     save numeric,
     limitation numeric,
-    required integer,
+    required numeric,
     alignment alignment,
     is_gang_addition boolean,
     alliance_id uuid,
